✨[feat]: BankManager 파일 생성 및 구현
