+++ BankManagerConsoleApp/BankManagerConsoleApp/Bank.swift
@@ -1,7 +1,8 @@
 //
-//  BankManager.swift
-//  Created by yagom.
-//  Copyright Donnie, Safari. All rights reserved.
+//  Bank.swift
+//  BankManagerConsoleApp
+//
+//  Created by Donnie, Safari on 2022/04/28.
 //
 
 import Foundation
@@ -15,30 +16,7 @@ struct Bank {
         self.bankClerkCount = bankClerkCount
     }
     
-    func open() {
-        printMenu()
-        guard let input = readLine() else { return }
-        
-        switch input {
-        case "1":
-            startWork()
-        case "2":
-            return
-        default:
-            print("잘못된 입력입니다.")
-            open()
-        }
-    }
-    
-    private func printMenu() {
-        print("""
-                1 : 은행개점
-                2 : 종료
-                입력 :
-                """, terminator: " ")
-    }
-    
-    private func startWork() {
+    func startWork() {
         let customers = Int.random(in: 10...30)
         for number in 1...customers {
             bankWaitingQueue.enqueue(Customer(numberTicket: number))
@@ -52,7 +30,7 @@ struct Bank {
             guard let customer = bankWaitingQueue.dequeue() else { return }
             DispatchQueue.global().async(group: group) {
                 bankWindows.wait()
-                bankClerk.processTask(for: customer)
+                self.bankClerk.processTask(for: customer)
                 bankWindows.signal()
             }
         }
@@ -61,7 +39,6 @@ struct Bank {
         let digit: Double = pow(10, 2)
         
         finishWork(customers, round(durationTime * digit) / digit)
-        open()
     }
     
     private func finishWork(_ customers: Int, _ totalWorkTime: Double) {
